script_engine: improve {,subdir} script loading
For now, they get loaded alphabetically no matter whether the script is
in the root directory or a subfolder, and no namespace collision checks
are performed.
This makes it hard to use original gamedata.db* and a mod where the
scripts are placed in subdirs, as then there'll be name clashes and
you can never tell which script was picked by the game (only the first
one get added to the map).
Instead, don't load scripts with the size of 0 bytes and crash on
namespace collision. IOW, now you can't have

dialogs.script
dialogs/dialogs.script

to avoid ambiguity, BUT if you want to override the original
dialogs.script, just replace it with a zero-byte stub and place
the right one to the dialogs/ folder.
The intended usage is that you supply gamedata.db_stubs with all
the scripts replaced with 0-byte stubs and then your actual gamedata
containing all the scripts you need, either inside subdirs or in the
root folder.

Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>