render: fix reading of level LODs
Structures in FLOD.h are shared with the game resources, so their layout
can't change. After the increased alignment for vectors, their size was
also increased leading to either LODs not properly working or crashes
when loading the game (with vanilla levels).
Replace Fvector2 with floats and add static assertions.
I rechecked the rest of FS::r(sizeof(x)) code and seems like we're fine
now.

Fixes: 91afadc74fc7 ("xrCore: use AVX2 for vectors and matrices")
Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>