alife: add "after_graph_load" global callback
Unlike the original engine, now the game graph is not ready when the
main menu appears, it gets build when the level loads. This means,
stuff like

_G.GGraph = game_graph()

in _G.script won't work anymore.
You can't currently catch when the graph is ready in the scripts.
To ease that, add a callback where you can put shortcuts like the
above one.

Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>