3rd_party: build xxHash statically
We only use a small piece of it (no streaming API etc.) and in a couple
places it's even embedded as all-inline.
Build it in the bundled mode. It only increases xrEngine's size by 2 Kb,
while the full dll was 39 :D
The only reason why I was building it dynamically is that before I
switched the whole project to Clang, MSVC was not able to statically
link ClanLTO objects.

Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>