xrGame: init game globals in parallel with UI creation
Run a task to initialize game globals in parallel with UI and main menu
creation, from OXR.

1. Move UI geom creation out of init_game_globals() to xrRender.
2. Create inventory shaders in UIGameCustom, not in init_game_globals().
3. Run a task for init_game_globals() to initialize them in parallel
   while the UI core and main menu are being initialized.

This shaves off some msecs when the window appears. Now it's 1-1.3
seconds from running the exe to the main menu, then 2 seconds for the
synchronization phase (the rest is almost instant), and it's ready to
play. I'm satisfied with the result, the only thing now worth
parallelizing is render.
The best fit would be to use a Flow Graph to describe all the
dependencies between phases. Phases must be thread-safe, both
calculation and rendering, and separated as much as possible.
Then we just run the graph and oneTBB tosses tasks between CPUs
and makes sure everything in synchronized.
Ah, dreams :D

Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>