build: LLVM 20, std=c++23
First of all, adjust paths to the LLVM root, so that the actual VS
property which allows to specify custom LLVM path will work. I don't
hardcode any custom version here as it requires specifying absolute
path to the compiler. That is not portable at all.

Next, switch the project to C++23. We don't use any particular
functionality from C++23 yet, but it's no problem and we can switch
already. The only thing that needed to be changed was Luabind's
class[1, 2, 3], i.e. commas inside [] operator, as in C++23 multi-dim
operator[] overload is introduced. So, the existing code only needed
to have an additional set of () after [], i.e. class[(1, 2, 3)].
This works just like pre-C++23 class[1, 2, 3].
BTW Visual Studio is cringe even here as MSVC doesn't currently support
full C++23. When using "stdcpp23" in vcxproj, it adds compiler flag
/std:c++23preview even when building with Clang, which is ignored by
it as it fully supports full C++23. So, in order to build in 23 mode
correctly with Clang, you also needs to manually add -std=c++23, VS
doesn't give a fuck >_<

Also, while here, remove all `pragma (optimize "s")`, it doesn't work
on Clang and I don't know why this was done. This engine is
LLVM-exclusive, so I don't care much if this changes MSVC behaviour
anyhow, but probably not.

Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>