xrCore: fix pausable timers and g_PauseMngr after recent changes
Currently, when the level is loaded and you open main menu and then
close, even after opening it, level game time keeps increasing, which
means timers are not really paused when opening the menu.
I don't know the exact reason, but that change from OXR fixes that.
When you register a paused timer (which happens e.g. on global Device
singleton creation), use function which will create a new static
instance and initialize pointer, former g_PauseMngr. This is needed
in order to make sure the pause manager is ready at the moment of
Device construction.
In rest places, just use the initialized pointer to avoid calling
getter each time.

Additionally, remove 2 obsolete leftovers:
IGamePersistent::CanBePaused(), since we only support single player
and this is always true; and m_paused from Environment, as it's not
set to true anywhere, so all the related code is dead.

Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>