xrGame: restore periodic Lua garbage collection
No idea why it was removed in OGSR. Push it to seqParallel to let it
work in the second thread.
The current settings is from OXR. The stub for LUA_GCTIMEOUT is here,
but as of now, I'm using upstream luajit2 from OpenResty, so it's not
supported. But even with LUA_GCSTEP it works just fine with no FPS
drops.

Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>