xrRender: don't clear rt_Accumulator and rt_Color each frame
In Anomaly, only rtPosition is cleared.
In OXR, the same, except that rt_Accumulator and rt_Color also get
cleared if MSAA is on.

Let's pick the first one, since the reference SSS are based on it.
I tested the game without clearing the acc and color at all both
with MSAA and without and don't see any issues.

When clearing the acc and color each frame, like it was done in OGSR,
I see black blinking time to time (each 5-10 seconds) when rendering
in several threads. Without it, it looks just fine.

The only two visible issues for now are:
1) When it's raining, sometimes (also 5-10 seconds) there are white
   blinks. If I add clearing rt_Accumulator to black after rendering
   the rain smap (like it's done in OXR), blinks get black, but stay.
   There was no such issue when rain was rendered within the main
   thread, so needs investigating.
2) Light shadow trails on walls from the weapon HUD and light noise
   on actor shadows on the HUD. This appeared after upgrading SSS
   Shadows to either 21 or 22, don't remember.
   Setting G_SSS_WEAPON_LENGTH in settings_screenspace_SSS.h makes
   this disappear, BUT in Anomaly, there's no such issue, and that
   default 0.09f was not chosen "just because".
   Dunno the exact reason as well, will see with SSS 23, where
   motion vectors are now per-object, not only for the HUD.

The rest is stable and nice (except 3090 performance lol).

Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>