xrEngine: port object scheduler logic from https://github.com/eagleivg/xray-16
In two words: instead of calculating some crazy stuff like time windows,
increase the budget in case of low number of object processed this frame
etc., it relies on the total time of the last frame and picks a part of
it, by default 1/2.
IOW, if we have 50 FPS, each frame takes 20 msecs, so the next frame,
the scheduler will have a budget of 10 msecs. If we have 144 FPS, like
my monitor, 3 ms will be taken (144 FPS = 6 msec/frame), which means
less objects will be updated, BUT it's not a problem at all since we
have more FPS!
This approach, while not affecting on FPS anyhow when the main
bottleneck is the GPU, greatly improves general smoothness. Previously,
when I was looking at the sky and nothing besides the skybox was drawn,
I still had like 110-120 FPS despite almost zero load on the GPU. Now
I have full 144 thanks to this.

The second part is that we now manually stop Lua GC after each
schedule_Update(). This is needed because Lua runs GC on each
allocation, but we have our own GC pass once per frame, so these
additional passes will only hurt.
To not add more includes and dependencies to xrEngine, I made this as
an IGame_Level method, and xrGame/Level.cpp defines it right next to
its GC-related stuff.

From my side, I replaced OGSR's `rs_sheduler_max`, which sets the top
limit of objects that can be processed each frame, to
`rs_scheduler_part`, which sets the percentage of the total frame time
that can be taken to update objects. By default, it's 0.5 like in the
original implementation (link above), but I realize that on different
systems, different values can provide best experience.
The range is 0.1-1.0, which means from 1/10 of frame time to the whole
frame time.

Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>