xrEngine: fix new scheduling algo
Ooops, I didn't notice we need to override the flag for collecting
the essential frame stats, otherwise we always had 0 there for
FrameTOTAL and the scheduler always took 1 ms as its budget...

+ calling LUA_GCSTOP after every object update is redundant and makes no
sense. This call stops the GC until it's restarted or run manually.
Just call it manually after the seqFrameMT GC pass.

Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>