3rd_party: remove LZO
Now that we have a better archive format which doesn't use LZO and
doesn't compress stuff per-file, we can deprecate and remove LZO.
The only place there it's still used is savegames. Replace LZO with
ZSTD with low compression level (1) there. A save made on vanilla
right after exiting Sidorovich's bunker was usually about 920-950
Kb, now is's 610 and the game saves and loads even faster lol.
I leave XDB format support, but now there's an explicit assertion
that if a file is LZO-compressed in the archive, this is not
supported and the game will exit.

Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>