xrGame: also call "after_objs_load" callback when starting a new game
Currently, it's called only when loading a save, making it impossible
to use for initializing global vars in scripts.
Call it when creating a new game, too, but pass (nullptr, 0) as
arguments. The script can then check it with r_eof() and realize it's
a new game, but still initialize stuff.

Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>