xrCore: make LzHuf thread-safe
This piece of code, originally written in 1988, has a fuckload of static
variables/arrays, total > 16 Kb size. Most important, they are not
anyhow protected from concurrent access from several threads.
Encapsulate them in a class and make the functions currently used by
the engine do xr_new<>() -> comp/decomp() -> xr_delete() on the fly,
leaving nothing in the global space.
I could place that LzHuf on the stack, but it's really huge, while
mimalloc is fast, so let it be this way. It's not used very often
after all.

Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>