xrCore: switch xr_*_{map,set} to absl:: equivalents
"Almost" drop-in replacements. The most visible difference is that
iterators are invalidated on insert or erase.
So this needed a couple adjustments in the xrGame code. In one place,
std::map is left -- Car.h for wheels, as wheels themselves are allocated
within the map and then pointers used everywhere throughout the car
code. As a result, it's really difficult to track everything down out
there.
The rest went fine.

Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>