xrRender: retire FixedMap
We have Abseil.
For static and dynamic geometry maps, we can use unordered map, as fast
pass lookup is important, while the order not at all. Instead of
converting them to a vector and then sorting it by SSA, convert them to
xr_map, so that they will already be sorted.
For vectors of elements inside those maps, use xr_multimap instead of
std::vector + sort(), as we know the sorting rule in advance (SSA).
For front-to-back etc. maps we can use xr_multimap, which will be sorted
on the fly, just like FixedMap, but better.
Less allocations, less iterations, smaller element size (esp. MatrixS).

Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>