3rd_party/luabind: I Won't See You Tonight Pt 2 (WIP)
Despite my first negative experience, I decided to move forth and got
really good progress.

First of all, I didn't manage to reduce the binary size for now (this
is due to that Sol effectively expands all its bindings to direct Lua C
API calls), but all the problems with compilation and linking time are
resolved. I had to switch from FatLTO to ThinLTO, which performs final
linking in several threads without hurting the optimization, and that's
it. Now, instead of 40 minutes, the final exe links in ~2 minutes.

Most important in this commit:

* as I mentioned, switch both 3rd_party and engine libraries to
  -flto=thin to get reasonable compilation and link times;

* enable LUAJIT_ENABLE_LUA52COMPAT unconditionally when building LuaJIT
  to better integrate with Sol features. For containers like vectors,
  maps etc, Sol generates __{,i}pairs() metamethods, which work only
  in Lua 5.2+. This definition doesn't break anything in the game;

* declare Sol traits with exported classes' parents using
  XR_SOL_BASE_CLASSES() -- same crap as SOL_BASE_CLASSES(), but
  generates all the lists automatically thanks to our RTTI with
  parent reflection.
  This is needed to perform proper casts in Lua from derived to base
  classes, not only pointers, but also even unique_ptr<> and friends;

* pass sol::state_view& directly instead of lua_State* to all
  script_register() methods to optimize the code a bit more;

* implement Lua-defined ops and priv for:
  * main menu client class;
  * all server classes from class_registrator.script. This includes
    stuff like STATE_Read/Write and all that was previously
    implemented as Luabind overridable methods;
  * camera effector (used in sr_aes_deadzone.script and
    ogsm_effector.script);
  * property_evaluator, action_base, everything from modules.script --
    used in NPC client schemes;

* also implement both client and server factory functions to be able
  to register client classes from Lua -- previously, Luabind custom
  cast functions were used for that;

* main menu, credits, options, load/save windows, ui_numpad --
  now fully work!
  (no compat with vanilla scripts, I'll push the rewritten versions to
   reaper-build once the whole game works stable);

* implement 3 different constructors for vector{2,3,4} to avoid calling
  :set() every time -- will save 1 cross-language call each time a
  vector is used;

* sprinkle XR_NOVTABLE (thx to OpenXRay) to reduce the binary size
  a bit (there are more places to address, but I need a fully
  working game first);

* stop initializing Luabind at all at this point;

* change the function which registers the whole chain of exported
  classes (using imdex::typelist) from open-coding all the calls to
  storing them in an array and calling in a loop -- reduces binary
  size as well;

* convert alife.ltx::start_game_callback and all OGSR global callbacks
  to sol::function.

There are still around 280 "luabind" keywords in xrGame -- will be
eventually converted when rewriting and testing the corresponding
classes.

The only class not switched to the new scheme is object_binder -- client
objects start spawning at the final game loading stage, I'd like to test
server stuff and everything that loads before them first.
In order to be able to do that, I still need to rewrite around 250 Lua
"classes" from Luabind custom scheme to pure Lua. Most of them
(excluding the abovementioned object_binder) are pure Lua objects which
aren't used anyhow in the engine, so this conversion will be straight
and smooth.

Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>