xrEngine: drop all obsolete calling convention specifiers
__cdecl, __stdcall, __fastcall are no-op on x64, so just drop it.
__vectorcall is pathetic just like the original MS x64 calling
convention; the only convention we care about is SysV, for which we
already have `XR_SYSV` macro and which I'll be using more and more
often.
I mean, it _sorta_ introduces function coloring (calling a SysV function
from an MS function can induce a bit of overhead, but calling an MS
function from a SysV function can not), so when I rewrite the core code
to be coroutine-based, I'll sprinkle XR_SYSV in lots of places right
away, so that we could use it without thinking of any side effects.

Misc: also drop _BCL since it's an empty macro for years already.

Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>