xrGame: fix extremely rare crash when loading string tables
Half a year passed since the tagged commit and I caught this crash for
the first time ever...
Anyway, ParseLine() is not thread-safe since it can insert entries to
the keybinding map (to update string tables when rebinding a key).
It's enough to wrap this insertion with the same mutex that protects
string tables themselves, as this happens before we take it to insert
a string.

Fixes: 2a5834474b19 ("xrGame: load string tables in parallel")
Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>