xrCore: string API global overhaul (+ smem_container)
Key points:
* xr_strcmp(), xr_strlen() are now fully-C++, fully constexpr and fully
  transparent (with all sorts of overloads: gsl::czstring aka
  `const char*`, absl::string_view, shared_str, xr_string) and are
  implemented using the corresponding string_view's methods.
* few usages of strncmp() are replaced with string_view{}.starts_with(),
  as it's the same, but also fully-constexpr-C++-blah;
* _stricmp() is implemented in C++ as constexpr xr::strcasecmp() using
  open-coded constexpr xr::tolower();
* all string comparison functions and operators are now three-way (aka
  <=> aka spaceship) -- almost the same as < 0, 0, > 0 (what memcmp()
  and strcmp() return), but strictly typed. You can still compare the
  result with zero or use std::is_*() family like I did in this commit
  (std::is_eq(xr_strcmp()) is the same as xr_strcmp() == 0);
* shared_str always stores its length, so we can construct string_views
  from it faster (omitting implicit string length calculation);
* xr_strlen now returns gsl::index (ptrdiff_t) as I started to slowly
  migrate my code and my mindset to signed maths;
* replace all plain C strcmp() and strlen() calls with ours;
* but the main change is:

str_container and smem_container are fully rewritten to be completely
lockless, thread-safe, and as fast as possible.

They are implemented as a lockless MPMC queue of unordered maps. This
means that on a CPU with 24 HW threads we may end up (in the worst case)
with 24 shared_str maps and 24 smem maps with some amount of duplicates.
But I can't afford mutexes in constructors (lol) and in general this
turned to be just faster than what previously was here.
smem_container was just a vector and str_container was a homemade
C-style hashmap with only 18 bits and a fixed array of 256k pointers
(so that 1 Mb of memory was wasted right away). I'm not sure how fast
std::unordered_map would work here, but absl::flat_hash_map *is* fast.
We still calculate hash for each string and each smem once (using xxH),
but it in fact helps us later -- instead of string (per-char)
comparison, we can just compare hashes (same for smem).

I'm satisfied with the resulting performance, memory usage (previously,
the "memory economy" line was saying something like 12 Mb, now it's from
7 to 9, but previous stats weren't completely honest, and as I said,
there will be some string duplication, but being lock-free is more
important here), and the code. We can now remove that conditional
disabling of smem when we have > 32 Gb of RAM -- new smem_container
works as fast as not having it at all, while even with a few NPC models
loaded it already saves 30 Mb (which also means 30 Mb of allocating,
copying, and freeing).

Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>