xrCDB: strictly-typed X-Ray pt II
Nothing outstanding in particular, a bit more active usage of std::span
and f32 and that's mostly it. There was a couple looks-like-a-real-bug,
but seems like it's been working previously somehow.
No changes in .exe size made me mad 'cause I tried so hard to optimize
things, but LLVM just does that for me and better than me T.T

Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>