3rd_party: move concurrentqueue, hwloc, and oneTBB to Src/concurrency/
We now have 3 submodules related to concurrency and this number will
only grow (with the next commit actually, but not only).
Create a folder 'concurrency' and move the submodules there.
No functional changes. I also didn't reorganize the projects in
.slnx as we build only hwloc and oneTBB, but the latter will be
removed sooner or [rather] later and the replacement for it (TMC,
ASIO et al) are either header-only or .ipp-based.

Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>