use BitmapObjectPool for string maps (#1)
[comment from Alex]

BitmapObjectPool is what we now should use instead of an MPMC queue
for lockless per-thread pools of objects.
It's way more lightweight, way faster, LIFO performs way better in
such cases -- string economy stat is now 11-12 Mb, while using the
queue it was 8, and without the pool and with only one map it was
around 14, which means we now have way less string duplicates
between maps (with only a couple Mb less than the reference).

There are 3 such places left in the engine: SquashFS readers, random
gens, and the smem container -- will be converted in the subsequent
commits.

Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>