xrCore: use BitmapObjectPool for SquashFS readers
Finish replacing ConcurrentQueue with BitmapObjectPool by rewriting
SquashFS support.
The pool recently got support to pass arbitrary parameters to
acquire_scoped(), use it to pass references to the archives
(archives are stored in a vector, so we can't promise their
addresses won't change).
Refactor my own code which I wrote half a year ago to be more C++
-- my experience has evolved a lot lol. Use std::unique_ptr<> and
absl::WrapUnique() for everything we can to delegate resource
management to the constructors and destructors.

Misc changes:
* pass SQFS_FILE_OPEN_NO_CHARSET_XFRM when creating a SquashFS file
  to resolve issues with non-ASCII paths (thx OGSR community);
* declare the default deleter for CStreamReader the same way as for
  IReader to be able to safely use absl::WrapUnique();
* wrap squashfs-tools-ng's C headers into `sqfs` namespace for
  better isolation;
* use the updated acquire_scoped() to automatically reserve space
  in string pool maps (xrstring.cpp).

Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>