misc: replace own xr_bitmap<> with plf::bitset<>
It was written a while ago only to replace the awful std::bitset<>
from MS STL. It doesn't support > 64 bits and the only thing it
does good -- searches the first zero bit using the MS intrinsic.
Replace it with plf::bitset<>, which already uses intrinsics,
but provides +/- the same functionality as regular std::bitset<>,
even richier (allows to use types other than size_t for the
underlying storage etc.).
Also, replace the MS intrinsic in agent_enemy_manager.cpp with
the one from std (resolves to the Clang built-in with libc++).

Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>