xrRender: fix an exit-time crash/UB
occq_destroy() gets called manually when uninitializing the
"device", but also from the destructor of CRender/RImplementation.
Since RImplementation is static, its destructor is exit-time.
We also have a mutex in xrCore/log.cpp, which protects concurrent
accesses to the log iostream. It's also static, so its destructor
is -- tadah -- exit-time. And it gets called *before* the
destructor of CRender.
As a result, the log and its mutex gets destroyed and then, a bit
later, occq_destroy() tries to write a couple messages to the log.
On LLVM libc++, this leads to an exit-time crash, which can be
caught only under VS debugger or Tracy. MS STL seems to get away
with this somehow, but this is a real bug anyway.

(BTW it's also a zero-day bug and was present even before I forked
 the upstream OGSR)

Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>